bugs fixed
git-svn-id: http://konoha.googlecode.com/svn/trunk/konoha@131 11e8e06f-943d-0410-b0b4-4be84c385469

+++ src/main/struct.c
@@ -2353,22 +2353,22 @@ static void Assurance_checkin(CTX ctx, knh_sfp_t *sfp, Object *o)
 	knh_Assurance_t *g = (knh_Assurance_t*)o;
 	g->aid = uid++;
 	g->stime = (knh_getTimeMilliSecond() / 1000);
-	KNH_SYSLOG_(ctx, sfp, LOG_NOTICE, "ac", "CHECKIN", "id=%d, case='%s'", (int)g->aid, S_tochar(g->msg));
+	KNH_SYSLOG_(ctx, sfp, LOG_NOTICE, "ac", "CHECKIN", "id=%d, case='%s'", g->aid, S_tochar(g->msg));
 }
 
 static void Assurance_checkout(CTX ctx, Object *o, int isFailed)
 {
 	knh_Assurance_t *g = (knh_Assurance_t*)o;
 	if(isFailed) {
-		KNH_SYSLOG_(ctx, NULL, LOG_WARNING, "ac", "FAILED", "id=%d, case='%s'", (int)g->aid, S_tochar(g->msg));
+		KNH_SYSLOG_(ctx, NULL, LOG_WARNING, "ac", "FAILED", "id=%d, case='%s'", g->aid, S_tochar(g->msg));
 	}
 	else {
 		knh_intptr_t t = (knh_getTimeMilliSecond() / 1000) - g->stime;
 		if(t > 1) {
-			KNH_SYSLOG_(ctx, NULL, LOG_NOTICE, "ac", "CHECKOUT", "id=%d, case='%s', time=%ds", (int)g->aid, S_tochar(g->msg), (int)t);
+			KNH_SYSLOG_(ctx, NULL, LOG_NOTICE, "ac", "CHECKOUT", "id=%d, case='%s', time=%ds", g->aid, S_tochar(g->msg), t);
 		}
 		else {
-			KNH_SYSLOG_(ctx, NULL, LOG_NOTICE, "ac", "CHECKOUT", "id=%d, case='%s'", (int)g->aid, S_tochar(g->msg));
+			KNH_SYSLOG_(ctx, NULL, LOG_NOTICE, "ac", "CHECKOUT", "id=%d, case='%s'", g->aid, S_tochar(g->msg));
 		}
 	}
 	if(knh_getUFILE() != NULL) {
